+++ src/memory/index.ts
@@ -5,3 +5,4 @@ export * from "./digest.js";
 export * from "./embed.js";
 export * from "./observe.js";
 export * from "./search.js";
+export * from "./lessons.js";